requirements.txt
sklearn~=0.0
pandas~=1.0.3
numpy~=1.18.4
opencv-python~=4.5.1.48
scikit-learn~=0.23.0
tqdm~=4.60.0
pillow~=8.2.0
albumentations~=0.5.2
numba~=0.53.1
tifffile~=2021.4.8
rasterio~=1.2.3
matplotlib~=3.4.1

sklearn~=0.0
joblib~=0.14.1
wheel
pandas~=1.0.3
requests~=2.23.0
scipy~=1.4.1
numpy~=1.18.4
opencv-python~=4.5.1.48
torchvision~=0.9.1
timm~=0.4.5
scikit-learn~=0.23.0
pydicom~=2.1.2
tqdm~=4.60.0
pillow~=8.2.0
scikit-image~=0.18.1
datetime~=4.3
albumentations~=0.5.2
